All colors/options are now defined in the __init__.py file instead of being spread around the whole gui.
git-svn-id: https://svn.tribler.org/abc/branches/mainbranch@18058 001aeff7-3401-0410-a489-f7902fc005dd

+++ Tribler/Main/vwxGUI/home.py
@@ -149,11 +149,11 @@ class NewTorrentPanel(HomePanel):
     
     def OnNotify(self, subject, type, infohash):
         torrent = self.torrentdb.getTorrent(infohash, include_mypref=False)
-        
-        self.list.InsertStringItem(0, torrent['name'])
-        size = self.list.GetItemCount()
-        if size > 10:
-            self.list.DeleteItem(size-1)
+        if torrent:
+            self.list.InsertStringItem(0, torrent['name'])
+            size = self.list.GetItemCount()
+            if size > 10:
+                self.list.DeleteItem(size-1)
     
     def OnDoubleClick(self, event):
         selected = self.list.GetFirstSelected()
